[13] - Classes moved to a new file, small refactors and some changes from PR done
