+++ koans/Classes/ExampleClass.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace PhpKoans\Classes;
+
+class ExampleClass
+{
+    public string $exampleProperty = 'Hi, I\'m a property';
+
+    public function exampleMethod(): string
+    {
+        return 'Hi, I\'m a method';
+    }
+}
